+++ domain-ee/ee-ep-merge-app/end_to_end/test_resume_at_add_note.py
@@ -8,6 +8,8 @@ from src.python_src.api import JOB_RUNNER
 from src.python_src.schema.merge_job import JobState, MergeJob
 from src.python_src.service.job_store import JOB_STORE
 
+from .conftest import assert_job, reset_claim
+
 PENDING_CLAIM_ID = 10000
 EP400_WITH_DUPLICATE = 10001
 EP400_WITH_DIFFERENT_CONTENTION_TYPE_CODE = 10002
@@ -19,20 +21,6 @@ CLAIM_ID_ERROR_AT_GET_CLAIM_DETAILS = 5001
 CLAIM_ID_ERROR_AT_UPDATE_CONTENTIONS = 5005
 
 NOW = datetime.now()
-
-
-def assert_job(job_id, pending_claim_id, ep400_claim_id, expected_state: JobState, expected_error_state: JobState | None = None, expected_num_errors: int = 0):
-    job = JOB_STORE.get_merge_job(job_id)
-    assert job is not None
-    assert job.pending_claim_id == pending_claim_id
-    assert job.ep400_claim_id == ep400_claim_id
-    assert job.state == expected_state
-    assert job.error_state == expected_error_state
-
-    if expected_num_errors == 0:
-        assert job.messages is None
-    else:
-        assert len(job.messages) == expected_num_errors
 
 
 class TestSuccess:
@@ -47,7 +35,7 @@ class TestSuccess:
             pytest.param(PENDING_CLAIM_ID, EP400_WITH_MULTI_CONTENTION_NO_DUPLICATES, id='with with no duplicates'),
         ],
     )
-    async def test(self, pending_claim_id, ep400_claim_id):
+    async def test(self, pending_claim_id, ep400_claim_id, lifecycle_endpoint):
         job_id = uuid4()
         job = MergeJob(
             job_id=job_id,
@@ -60,7 +48,10 @@ class TestSuccess:
         JOB_STORE.submit_merge_job(job)
 
         await asyncio.get_event_loop().run_in_executor(None, JOB_RUNNER.resume_job, job)
-        assert_job(job_id, pending_claim_id, ep400_claim_id, JobState.COMPLETED_SUCCESS)
+        try:
+            assert_job(job_id, pending_claim_id, ep400_claim_id, JobState.COMPLETED_SUCCESS)
+        finally:
+            await reset_claim(ep400_claim_id, lifecycle_endpoint)
 
 
 class TestError:
@@ -84,7 +75,7 @@ class TestError:
             ),
         ],
     )
-    async def test(self, pending_claim_id, ep400_claim_id, expected_error_state, expected_num_errors):
+    async def test(self, pending_claim_id, ep400_claim_id, expected_error_state, expected_num_errors, lifecycle_endpoint):
         job_id = uuid4()
         job = MergeJob(
             job_id=job_id,
@@ -97,4 +88,7 @@ class TestError:
         JOB_STORE.submit_merge_job(job)
 
         await asyncio.get_event_loop().run_in_executor(None, JOB_RUNNER.resume_job, job)
-        assert_job(job_id, pending_claim_id, ep400_claim_id, JobState.COMPLETED_ERROR, expected_error_state, expected_num_errors)
+        try:
+            assert_job(job_id, pending_claim_id, ep400_claim_id, JobState.COMPLETED_ERROR, expected_error_state, expected_num_errors)
+        finally:
+            await reset_claim(ep400_claim_id, lifecycle_endpoint)
